Removed react-table implementation and cleaned up material-table implementation
