+++ packages/core/package.json
@@ -30,7 +30,6 @@
     "@types/google-protobuf": "^3.7.2",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
-    "@types/react-table": "^7.0.16",
     "classnames": "^2.2.6",
     "clsx": "^1.1.0",
     "material-table": "^1.57.2",
@@ -41,9 +40,6 @@
     "react-dom": "^16.12.0",
     "react-helmet": "5.2.1",
     "react-router-dom": "^5.1.2",
-    "react-table": "^7.0.4",
-    "react-virtualized-auto-sizer": "^1.0.2",
-    "react-window": "^1.8.5",
     "recompose": "0.30.0"
   },
   "devDependencies": {
