+++ packages/core/src/components/Table/Table.tsx
@@ -14,170 +14,172 @@
  * limitations under the License.
  */
 
-import React, { FC } from 'react';
-import { useTable, useGlobalFilter, useSortBy } from 'react-table';
-import type { Column } from 'react-table';
-import classNames from 'classnames';
+import React, { FC, forwardRef } from 'react';
+import MTable, {
+  MTableCell,
+  MTableHeader,
+  MTableToolbar,
+  MaterialTableProps,
+  Options,
+  Column,
+} from 'material-table';
+import { BackstageTheme } from '@backstage/theme';
+import { makeStyles } from '@material-ui/core';
 
+// Material-table is not using the standard icons available in in material-ui. https://github.com/mbrn/material-table/issues/51
 import {
-  Table as MUITable,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  TableSortLabel,
-  TextField,
-  makeStyles,
-} from '@material-ui/core';
+  AddBox,
+  ArrowUpward,
+  Check,
+  ChevronLeft,
+  ChevronRight,
+  Clear,
+  DeleteOutline,
+  Edit,
+  FilterList,
+  FirstPage,
+  LastPage,
+  Remove,
+  SaveAlt,
+  Search,
+  ViewColumn,
+} from '@material-ui/icons';
 
-const useHighlightCellStyles = makeStyles(theme => ({
-  root: {
-    color: theme.palette.grey[900],
-  }
-}));
+const tableIcons = {
+  Add: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <AddBox {...props} ref={ref} />
+  )),
+  Check: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Check {...props} ref={ref} />
+  )),
+  Clear: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Clear {...props} ref={ref} />
+  )),
+  Delete: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <DeleteOutline {...props} ref={ref} />
+  )),
+  DetailPanel: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <ChevronRight {...props} ref={ref} />
+  )),
+  Edit: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Edit {...props} ref={ref} />
+  )),
+  Export: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <SaveAlt {...props} ref={ref} />
+  )),
+  Filter: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <FilterList {...props} ref={ref} />
+  )),
+  FirstPage: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <FirstPage {...props} ref={ref} />
+  )),
+  LastPage: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <LastPage {...props} ref={ref} />
+  )),
+  NextPage: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <ChevronRight {...props} ref={ref} />
+  )),
+  PreviousPage: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <ChevronLeft {...props} ref={ref} />
+  )),
+  ResetSearch: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Clear {...props} ref={ref} />
+  )),
+  Search: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Search {...props} ref={ref} />
+  )),
+  SortArrow: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <ArrowUpward {...props} ref={ref} />
+  )),
+  ThirdStateCheck: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <Remove {...props} ref={ref} />
+  )),
+  ViewColumn: forwardRef((props, ref: React.Ref<SVGSVGElement>) => (
+    <ViewColumn {...props} ref={ref} />
+  )),
+};
 
-const useCellStyles = makeStyles(theme => ({
+const useCellStyles = makeStyles<typeof BackstageTheme>(theme => ({
   root: {
-    padding: theme.spacing(1, 2, 1, 2),
-  }
-}));
-
-const useHeadStyles = makeStyles(theme => ({
-  root: {
-    '&:first-child': {
-      fontWeight: 'bold',
-    },
-    borderBottom: `1px solid ${theme.palette.grey[500]}`,
-    borderTop: `1px solid ${theme.palette.grey[500]}`,
-    textTransform: 'uppercase',
+    color: theme.palette.grey[500],
+    padding: theme.spacing(0, 2, 0, 2.5),
+    height: '56px',
   },
 }));
 
-const useSortLabelStyles = makeStyles(theme => ({
-  icon: {
-    margin: theme.spacing(0, 0.5, 0, 0.5),
-  }
+const useHeaderStyles = makeStyles<typeof BackstageTheme>(theme => ({
+  header: {
+    padding: theme.spacing(1, 2, 1, 2.5),
+    borderTop: `1px solid ${theme.palette.grey.A100}`,
+    borderBottom: `1px solid ${theme.palette.grey.A100}`,
+    color: theme.palette.textSubtle,
+    fontWeight: 'bold',
+    position: 'static',
+  },
 }));
 
-type GlobalFilterProps = {
-  setGlobalFilter: any;
-  preGlobalFilteredRows: any;
-  globalFilter: any;
-};
-
-const GlobalFilter: FC<GlobalFilterProps> = ({
-  setGlobalFilter,
-  preGlobalFilteredRows,
-  globalFilter,
-}) => {
-  const count = preGlobalFilteredRows.length;
-
-  return (
-    <TextField
-      value={globalFilter || ''}
-      placeholder={`${count} records...`}
-      label="Search"
-      onChange={e => {
-        setGlobalFilter(e.target.value || undefined);
-      }}
-    />
-  );
-};
-
-export type ColumnProps = {
-  align?: string;
-  highlight?: boolean;
-};
-
-type TableProps = {
-  data: Array<any>;
-  columns: Array<Column<{}>> & ColumnProps;
-  showFilter?: boolean;
-};
-
-const Table: FC<TableProps> = ({ columns, data, showFilter = true }) => {
-  const headerClasses = useHeadStyles();
-  const sortLabelClasses = useSortLabelStyles();
-  const cellClasses = useCellStyles();
-  const highlightClasses = useHighlightCellStyles();
-
-  const {
-    getTableProps,
-    headerGroups,
-    rows,
-    prepareRow,
-    preGlobalFilteredRows,
-    setGlobalFilter,
-    state,
-  } = useTable(
-    {
-      columns,
-      data,
+const useToolbarStyles = makeStyles<typeof BackstageTheme>(theme => ({
+  root: {
+    padding: theme.spacing(3, 0, 2.5, 2.5),
+  },
+  title: {
+    '& > h6': {
+      fontWeight: 'bold',
     },
-    useGlobalFilter,
-    useSortBy,
-  );
+  },
+}));
+
+const convertColumns = columns => {
+  return columns.map(column => {
+    const headerStyle: React.CSSProperties = {};
+    const cellStyle: React.CSSProperties = {};
+
+    if (column.highlight) {
+      headerStyle.color = '#000000';
+      cellStyle.fontWeight = 'bold';
+    }
+
+    return {
+      ...column,
+      headerStyle,
+      cellStyle,
+    };
+  });
+};
+
+export interface TableColumn extends Column<{}> {
+  highlight?: boolean;
+}
+
+export interface TableProps extends MaterialTableProps<{}> {
+  columns: TableColumn[];
+}
+
+const Table: FC<TableProps> = ({ columns, options, ...props }) => {
+  const cellClasses = useCellStyles();
+  const headerClasses = useHeaderStyles();
+  const toolbarClasses = useToolbarStyles();
+
+  const MTColumns = convertColumns(columns);
+
+  const defaultOptions: Options = {
+    headerStyle: {
+      textTransform: 'uppercase',
+    },
+  };
 
   return (
-    <>
-      {showFilter ? (
-        <GlobalFilter
-          preGlobalFilteredRows={preGlobalFilteredRows}
-          setGlobalFilter={setGlobalFilter}
-          globalFilter={state.globalFilter}
-        />
-      ) : null}
-      <TableContainer>
-        <MUITable {...getTableProps()}>
-          <TableHead>
-            {headerGroups.map(headerGroup => (
-              <TableRow {...headerGroup.getHeaderGroupProps()}>
-                {headerGroup.headers.map(column => (
-                  <TableCell
-                    className={classNames(
-                      headerClasses.root,
-                      cellClasses.root,
-                      {[highlightClasses.root]: column.highlight}
-                    )}
-                    align={column.align === 'right' ? 'right' : 'left'}
-                    {...column.getHeaderProps(column.getSortByToggleProps())}
-                  >
-                    <TableSortLabel
-                      classes={sortLabelClasses}
-                      active={column.isSorted}
-                      direction={column.isSortedDesc ? 'desc' : 'asc'}
-                    >
-                      {column.render('Header')}
-                    </TableSortLabel>
-                  </TableCell>
-                ))}
-              </TableRow>
-            ))}
-          </TableHead>
-          <TableBody>
-            {rows.map((row, _i) => {
-              prepareRow(row);
-              return (
-                <TableRow {...row.getRowProps()}>
-                  {row.cells.map(cell => {
-                    return (
-                      <TableCell
-                        className={cellClasses.root}
-                        align={cell.column.align === 'right' ? 'right' : 'left'}
-                        {...cell.getCellProps()}
-                      >
-                        {cell.render('Cell')}
-                      </TableCell>
-                    );
-                  })}
-                </TableRow>
-              );
-            })}
-          </TableBody>
-        </MUITable>
-      </TableContainer>
-    </>
+    <MTable
+      components={{
+        Cell: props => <MTableCell className={cellClasses.root} {...props} />,
+        Header: props => <MTableHeader classes={headerClasses} {...props} />,
+        Toolbar: props => <MTableToolbar classes={toolbarClasses} {...props} />,
+      }}
+      options={{ ...defaultOptions, ...options }}
+      columns={MTColumns}
+      icons={tableIcons}
+      {...props}
+    />
   );
 };
